Settting up a basic map
In this commit I generated three scripts;
Noise, MapGenerator, MapDisplay and MapGenereatorEditor. These scripts allowed me to make and display a perlin noise map. The map can get generated by the "Generate-button" in the MapGenerator game object or by changing by the values of the image (when auto update is activated)